2 morphologies acting very similarly
The model with the reconstructed morphology acted very differently then
the other model/physiological firing pattern. I've found that in the
morphology there was a 0 as a diameter of a 'segment' of the soma. The
diam was not 0, but because of it Neuron had problems.
I've also added a line to the template file under the biophys() proc:
forsec all if (ismembrane("ar")) { m0_ar = 0 }